+++ GpsMidGraph/de/ueller/midlet/gps/routing/ConnectionWithNode.java
@@ -16,7 +16,7 @@ public class ConnectionWithNode {
 	public byte wayConStartBearing=0;
 	public float wayDistanceToNext=Float.MAX_VALUE;
 	/** time in 1/10 secs to next connection */
-	public float wayDurationToNext=0;
+	public int wayDurationToNext=0;
 	
 	public ConnectionWithNode(RouteNode n,Connection c){
 		this.to=n;
